Example code for Default Arguments
A default argument is an argument that assumes a default value if a value is not provided in the function call in that argument. 
The most useful form is to specify a default value for one or more arguments. This creates a function that can be called with fewer arguments than it is defined to allow.