Example Code For Keyword Arguments
Keyword arguments  relate to the function calls. When you use keyword arguments in a function call, the caller identifies the arguments by the parameter name.
This allows you to skip arguments or place them out of order because the Python the interpreter is able to use the keywords provided to match the values with parameters.